Made the tool argument in the constructor of RegressionTest and optional argument by giving it a default value.
If no argument for tool is given in the constructor of RegressionTest,
then the class will look through the PATH variable of the machine searching
for omc and dymola and execute the tests for all of the found simulators.

Nicer output formatting to make results of test output more readable.