Tracked down the error that occured when cleaning up folders under Windows.
To run the omc on the model to be tested I changed the working directory into the result folder, but I never changed it back.
So I was trying to delete the result folders in the cleanup method although it was still working directory. Under Linux no issue,
but failed under Windows.