Fixed another relative path issue, when package folder is given as relative
Then a new relative path must be computed but with respect to the target
test folder.

This is important when the generator is used to generate test for a
repository, in which no absolute paths shall be present.

If paths give are absolute, then absolute paths are also used in the
generated file.